pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.160.0",
    "google-generativeai>=0.8.4",
    "plotly>=6.0.0",
    "streamlit>=1.42.0",
]
